requirements.txt
pyTelegramBotAPI==4.15.4
google-generativeai>=0.5.0